Move listener to TownyBlockListener
